fixes as per review
