+++ src/components/ApplicationDashboard/ApplicationDashboard.tsx
@@ -15,6 +15,10 @@ import GridCol from "@govuk-react/grid-col";
 import styled from "styled-components";
 import { ListFundingApplicationsQuery } from "../../API";
 
+import { daysLeft, friendlyDate } from "../../lib/dateandtime";
+
+import P from "@govuk-react/paragraph";
+
 export const ApplicationContainer = styled.div`
     background: ${GREY_3};
     padding: 20px;
@@ -42,6 +46,20 @@ export const ApplicationContainerTimeline = styled.div`
     text-align: center;
 `;
 
+interface Typed {
+    time?: number;
+    prefix?: string;
+}
+
+function showDaysLeft(closeDate: string | null) {
+    const result: Typed = { ...daysLeft(closeDate) };
+    return (
+        <>
+            <H4 mb={1}>{result.time}</H4>
+            <P mb={1}>{result.prefix}</P>
+        </>
+    );
+}
 interface Props extends RouteComponentProps {
     applications?: ListFundingApplicationsQuery; //change when not mocked
 }
@@ -52,64 +70,20 @@ export const ApplicationDashboard: FC<Props> = props => {
         props.applications.listFundingApplications &&
         props.applications.listFundingApplications.items;
 
-    function daysLeft(date: string | null) {
-        if (!date) {
-            return <div> No close date</div>;
-        }
-
-        const closeDate = new Date(date);
-        const closeDateTime = closeDate.getTime();
-        const today = new Date();
-        const timeToday = today.getTime();
-        const timeleftBig = closeDateTime - timeToday;
-        const timeLeft = Math.floor(timeleftBig / 1000 / 60 / 60);
-
-        let timeToShow = timeLeft;
-        let prefixToShow = "Hours";
-
-        if (timeLeft > 24 && timeLeft < 48) {
-            timeToShow = Math.floor(timeLeft / 24);
-            prefixToShow = "Day left";
-        } else if (timeLeft > 47) {
-            timeToShow = Math.floor(timeLeft / 24);
-            prefixToShow = "Days left";
-        } else if (timeLeft === 1) {
-            prefixToShow = "Hour left";
-        } else if (timeLeft < 0) {
-            timeToShow = 0;
-            prefixToShow = "Ended";
-        }
-
-        return (
-            <>
-                <H4 mb={0}>{timeToShow}</H4>
-                <span>{prefixToShow}</span>
-            </>
-        );
-    }
-
-    function friendlyDate(date: string | null) {
-        return (
-            date &&
-            new Date(date).toLocaleDateString("en-GB", {
-                day: "numeric",
-                month: "long"
-            })
-        );
-    }
-
     return (
         <>
             <Title>Dashboard</Title>
             <ApplicationContainer>
                 <H4 mb={1}> Applications </H4>
 
-                <span>In progress ({applications && applications.length})</span>
+                <span>
+                    In progress ({applications ? applications.length : 0})
+                </span>
 
                 {applications &&
                     applications.map(application => {
                         if (!application) {
-                            return <div> Loading...</div>;
+                            return null;
                         }
                         return (
                             <ApplicationContainerItem key={application.id}>
@@ -121,24 +95,26 @@ export const ApplicationDashboard: FC<Props> = props => {
                                         >
                                             {application.opportunityName}
                                         </Link>
-                                        <div>
+
+                                        <GridRow>
                                             Application number:
                                             {application.id}
-                                        </div>
-                                        <div>
+                                        </GridRow>
+
+                                        <GridRow>
                                             Opportunity:
                                             {application.id}
-                                        </div>
+                                        </GridRow>
                                     </GridCol>
                                     <GridCol setWidth="25%">
                                         <ApplicationContainerTimeline>
-                                            {daysLeft(application.closeDate)}
-                                            <div>
-                                                Deadline
-                                                {friendlyDate(
-                                                    application.closeDate
-                                                )}
-                                            </div>
+                                            {showDaysLeft(
+                                                application.closeDate
+                                            )}
+                                            Deadline{" "}
+                                            {friendlyDate(
+                                                application.closeDate
+                                            )}
                                         </ApplicationContainerTimeline>
                                     </GridCol>
                                     <GridCol setWidth="25%">
